+++ packages/eslint-plugin/src/rules/indent-new-do-not-use/index.ts
@@ -1382,6 +1382,10 @@ export default createRule<Options, MessageIds>({
       },
 
       VariableDeclaration(node) {
+        if (node.declarations.length === 0) {
+          return;
+        }
+
         let variableIndent = Object.prototype.hasOwnProperty.call(
           options.VariableDeclarator,
           node.kind,
@@ -1588,7 +1592,7 @@ export default createRule<Options, MessageIds>({
     );
 
     // For each ignored node selector, set up a listener to collect it into the `ignoredNodes` set.
-    const ignoredNodes = new Set();
+    const ignoredNodes = new Set<TSESTree.Node>();
 
     /**
      * Ignores a node